+++ __mocks__/browser-tabs-lock/index.ts
@@ -1,0 +1,7 @@
+export const acquireLockMock = jest.fn();
+export const releaseLockMock = jest.fn();
+
+export default class FakeLock {
+  acquireLock = acquireLockMock;
+  releaseLock = releaseLockMock;
+}